Skiller ut kall til aktørreg i egen klasse for å gjøre koden mer testbar
