+++ src/main/java/no/nav/familie/ks/oppslag/aktør/AktørController.java
@@ -2,7 +2,6 @@ package no.nav.familie.ks.oppslag.aktør;
 
 import no.nav.familie.ks.oppslag.felles.MDCOperations;
 import no.nav.security.oidc.api.ProtectedWithClaims;
-import no.nav.security.oidc.api.Unprotected;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -15,16 +14,15 @@ import javax.validation.constraints.NotNull;
 @RequestMapping("/api/aktoer")
 public class AktørController {
 
-    AktørregisterClient aktørregisterClient;
+    private AktørService aktørService;
 
-    AktørController(AktørregisterClient aktørregisterClient) {
-        this.aktørregisterClient = aktørregisterClient;
+    AktørController(AktørService aktørService) {
+        this.aktørService = aktørService;
     }
 
     @GetMapping
-    @Unprotected
     public String getAktoerIdForPersonIdent(@NotNull @RequestParam(name = "ident") String personIdent) {
         MDCOperations.putCallId(); // FIXME: Midlertidig, bør erstattes med en interceptor
-        return aktørregisterClient.getAktørId(personIdent);
+        return aktørService.getAktørId(personIdent);
     }
 }
