+++ src/main/resources/application-dev.yaml
@@ -6,8 +6,8 @@ no.nav.security.oidc:
     cookie_name: localhost-idtoken
 
 SECURITYTOKENSERVICE_URL: https://localhost:8063/soap/SecurityTokenServiceProvider/
-STS_REST_URL: http://security-token-service.default.svc.nais.local
-AKTOERID_URL: https://localhost:8063/aktoerregister/api/v1 # TODO: Må finne ut hva denne 8063-appen er
+STS_REST_URL: http://security-token-service.default.svc.nais.local/rest/v1/sts/token
+AKTOERID_URL: #https://localhost:8084/aktoerregister/api/v1 Denne bør mockes ut
 PERSON_V3_URL: https://localhost:8063/soap/tpsws/ws/Person/v3
 
 CREDENTIAL_USERNAME: not-a-real-srvuser
